feat: use kprobes to capture dns packets
Currently we use uprobes attached to the VM's glibc to intercept calls
to `getaddrinfo`, `gethostbyname`, and other related function calls.
This works really well to tell what host the process is reaching out to
and allows to get information on the that process is reaching out.

However, this has the downside that anything that doesn't use these
functions or uses some other instance of glibc is missed. For example,
any CGO-less binary will not use glibc, containerized workloads bring
their own glibc that we aren't attached to or may use musl.

This instead uses kprobes that we already had attached to the relevant
udp sending/receiving kernel functions to look for likely DNS packets
and forward them into user space to parse and correlate with other
events that we observe, similarly to how we use the uprobe information
today.