fix: fix ip_send_skb probe (#17)
* fix: fix ip_send_skb probe

The sk_buff is in a different state when we probe ip_send_skb versus
skb_consume_udp. During skb_consume_udp the buffer's head and data
pointers point to the network and user data portion of the packet,
respectively.

In ip_send_skb, both head and data point to the network header.
Additionally some other fields in the skb_buff are not set, such as the
IP version.

Due to this, we were unable to capture sent DNS packets, and only were
recording incoming packets. This patch fixes that by properly handling
the skb_buff differences in ip_send_skb.

* Update bpf/counter.c

Co-authored-by: Josh De Winne <joris.dewinne@gmail.com>

---------

Co-authored-by: Josh De Winne <joris.dewinne@gmail.com>